Add more unit tests for ConversationTurnComponent
